Move utils file
